style(reslicecursor): setup image instead of volume creation for ResliceCursor example
